refactor: extract to domain/shared
