+++ src/app/users/create-user.ts
@@ -5,7 +5,8 @@ import {
   AccountsRepository,
   UsersRepository,
 } from "@services/mongoose"
-import { WalletCurrency, WalletType } from "@domain/wallets"
+import { WalletCurrency } from "@domain/shared"
+import { WalletType } from "@domain/wallets"
 import { TwilioClient } from "@services/twilio"
 
 export const createUser = async ({
