+++ src/app/wallets/add-invoice-for-wallet.ts
@@ -6,7 +6,8 @@ import { checkedtoCents } from "@domain/fiat"
 import { RateLimitConfig } from "@domain/rate-limit"
 import { RateLimiterExceededError } from "@domain/rate-limit/errors"
 import { WalletInvoiceFactory } from "@domain/wallet-invoices/wallet-invoice-factory"
-import { checkedToWalletId, WalletCurrency } from "@domain/wallets"
+import { checkedToWalletId } from "@domain/wallets"
+import { WalletCurrency } from "@domain/shared"
 import { DealerPriceService } from "@services/dealer-price"
 import { LndService } from "@services/lnd"
 import { WalletInvoicesRepository, WalletsRepository } from "@services/mongoose"
