+++ src/app/wallets/check-limit-helpers.ts
@@ -3,7 +3,7 @@ import { LimitsChecker } from "@domain/accounts"
 import { toSats } from "@domain/bitcoin"
 import { toCents } from "@domain/fiat"
 import { TwoFA, TwoFANewCodeNeededError } from "@domain/twoFA"
-import { WalletCurrency } from "@domain/wallets"
+import { WalletCurrency } from "@domain/shared"
 import { LedgerService } from "@services/ledger"
 import { addAttributesToCurrentSpan } from "@services/tracing"
 import { mapObj } from "@utils"
