+++ src/app/wallets/send-intraledger.ts
@@ -3,7 +3,8 @@ import { getCurrentPrice } from "@app/prices"
 import { PaymentSendStatus } from "@domain/bitcoin/lightning"
 import { InsufficientBalanceError, NotImplementedError } from "@domain/errors"
 import { DisplayCurrencyConverter } from "@domain/fiat/display-currency"
-import { PaymentInputValidator, WalletCurrency } from "@domain/wallets"
+import { PaymentInputValidator } from "@domain/wallets"
+import { WalletCurrency } from "@domain/shared"
 import { LedgerService } from "@services/ledger"
 import { LockService } from "@services/lock"
 import {
