+++ src/app/wallets/update-pending-invoices.ts
@@ -3,7 +3,8 @@ import { InvoiceNotFoundError } from "@domain/bitcoin/lightning"
 import { DealerPriceServiceError } from "@domain/dealer-price"
 import { CouldNotFindError } from "@domain/errors"
 import { DisplayCurrencyConverter } from "@domain/fiat/display-currency"
-import { DepositFeeCalculator, WalletCurrency } from "@domain/wallets"
+import { DepositFeeCalculator } from "@domain/wallets"
+import { WalletCurrency } from "@domain/shared"
 import { LockService } from "@services/lock"
 import { DealerPriceService } from "@services/dealer-price"
 import { LedgerService } from "@services/ledger"
