+++ src/debug/create-wallets.ts
@@ -14,7 +14,8 @@ import { BTC_NETWORK } from "@config"
 import { checkedToSats } from "@domain/bitcoin"
 import { checkedtoCents } from "@domain/fiat"
 import { checkedToAccountLevel } from "@domain/users"
-import { checkedToWalletId, WalletCurrency, WalletType } from "@domain/wallets"
+import { checkedToWalletId, WalletType } from "@domain/wallets"
+import { WalletCurrency } from "@domain/shared"
 import { createToken } from "@services/jwt"
 import { isUp } from "@services/lnd/health"
 import { params as unauthParams } from "@services/lnd/unauth"
