+++ src/domain/ledger/activity-checker.ts
@@ -1,6 +1,6 @@
 import { toSats } from "@domain/bitcoin"
 import { toCents } from "@domain/fiat"
-import { WalletCurrency } from "@domain/wallets"
+import { WalletCurrency } from "@domain/shared"
 
 const MS_PER_HOUR = (60 * 60 * 1000) as MilliSeconds
 const MS_PER_DAY = (24 * MS_PER_HOUR) as MilliSeconds
