+++ src/domain/shared/index.ts
@@ -1,0 +1,2 @@
+export * from "./primitives"
+export * from "./calculator"
