+++ src/domain/shared/index.types.d.ts
@@ -1,0 +1,7 @@
+type WalletCurrency =
+  typeof import("./primitives").WalletCurrency[keyof typeof import("./primitives").WalletCurrency]
+
+type PaymentAmount<T extends WalletCurrency> = {
+  currency: T
+  amount: bigint
+}
