+++ src/domain/shared/primitives.ts
@@ -1,0 +1,5 @@
+// TODO: think how to differentiate physical from synthetic USD
+export const WalletCurrency = {
+  Usd: "USD",
+  Btc: "BTC",
+} as const
