+++ src/domain/wallets/primitives.ts
@@ -4,9 +4,3 @@
 export const WalletType = {
   Checking: "checking",
 } as const
-
-// TODO: think how to differentiate physical from synthetic USD
-export const WalletCurrency = {
-  Usd: "USD",
-  Btc: "BTC",
-} as const
