+++ src/graphql/root/mutation/ln-invoice-fee-probe.ts
@@ -5,7 +5,7 @@ import LnPaymentRequest from "@graphql/types/scalar/ln-payment-request"
 import { Wallets } from "@app"
 import { mapError } from "@graphql/error-map"
 import { WalletsRepository } from "@services/mongoose"
-import { WalletCurrency } from "@domain/wallets"
+import { WalletCurrency } from "@domain/shared"
 
 const LnInvoiceFeeProbeInput = GT.Input({
   name: "LnInvoiceFeeProbeInput",
