+++ test/integration/02-user-wallet/02-receive-lightning.spec.ts
@@ -5,7 +5,8 @@ import { MEMO_SHARING_SATS_THRESHOLD } from "@config"
 import { toSats } from "@domain/bitcoin"
 import { defaultTimeToExpiryInSeconds } from "@domain/bitcoin/lightning/invoice-expiration"
 import { toCents } from "@domain/fiat"
-import { PaymentInitiationMethod, WalletCurrency } from "@domain/wallets"
+import { PaymentInitiationMethod } from "@domain/wallets"
+import { WalletCurrency } from "@domain/shared"
 import { DealerPriceService } from "@services/dealer-price"
 import { LedgerService } from "@services/ledger"
 import { TransactionsMetadataRepository } from "@services/ledger/services"
